requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
Pillow>=8.0.0
segmentation-models-pytorch>=0.2.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
